+++ CJBItemSpawner/release-notes.md
@@ -4,6 +4,7 @@
 ## 1.12.1
 * Updated for Stardew Valley 1.3.29.
 * Updated list of problematic items not visible by default.
+* Fixed all slingshots spawned as the basic slingshot.
 
 ## 1.11
 * Updated for Stardew Valley 1.3 (including multiplayer support).